octavo_eeprom_flasher/requirements.txt
adafruit-ampy
pyserial
